analysis: major overhaul of class architecture
Most methods are now static.  Instances of the object can still be
created and used, but are limited to the scope of an individual file,
as they should be.  Batch analysis is called statically, and it works
by instantiating a new analysis object for each file.  Further event
analysis methods and plotting are called statically on events passed in.