analysis: changes to event finding again
While loops hopefully make this robust.  There is a new threshold called 'startendthreshold' which thresholds the exact start and end.  End is no longer the beginning of the end, rather it's the end of the end.  Tweaked open pore current finding just a bit to exclude possible rise and fall times.